lcapi/requirements.txt
asgiref==3.6.0
astroid==2.12.13
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
Django==4.1.4
django-admin-rangefilter==0.9.0
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.2
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
filelock==3.8.2
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.11.4
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
pathspec==0.10.3
phonenumbers==8.13.4
phonenumberslite==8.13.4
Pillow==9.1.1
platformdirs==2.6.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
whitenoise==6.3.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
